Pour que ça fonctionne mieux avec |image_recadre{x,y,focus} il faut qu'à chaque fois que l'on modifie manuellement le centre d'intérêt, que ça modifie le timestamp du fichier source, de sorte que SPIP comprenne que l'image à changé et donc nécessite d'actualiser les vignettes de cette image.
git-svn-id: svn://zone.spip.org/spip-zone/_plugins_@93478 ac52e18a-acf5-0310-9fe8-c4428f23b10a

+++ centre_image/trunk/action/centre_image_forcer.php
@@ -4,11 +4,20 @@ if (!defined("_ECRIRE_INC_VERSION")) return;
 
 
 function action_centre_image_forcer() {
-		$md5 = md5($_GET["url"]);
-		$forcer = sous_repertoire(_DIR_IMG, "cache-centre-image");
+	$fichier = $_GET["url"];
+	// image uniquement présente dans _DIR_IMG
+	if (strpos(_DIR_RACINE . $fichier, _DIR_IMG) === 0) {
+		if (file_exists(_DIR_RACINE . $fichier)) {
+			$md5 = md5($fichier);
+			$forcer = sous_repertoire(_DIR_IMG, "cache-centre-image");
 
-		$fichier_json = "$forcer$md5.json";
-		$res = array("x" => $_GET["x"], "y" => $_GET["y"]);
+			$fichier_json = "$forcer$md5.json";
+			$res = array("x" => $_GET["x"], "y" => $_GET["y"]);
 
-		file_put_contents($fichier_json, json_encode($res,TRUE));
+			@touch(_DIR_RACINE . $fichier);
+			file_put_contents($fichier_json, json_encode($res, TRUE));
+		}
+	}
+
+
 }
